el8: Add HostKeyAlias option to console attach ssh command
HostKeyAlias is set ovirt-vmconsole to ssh options.  It works well on
el7 but not on el8.  Newer OpenSSH versions have stricter HostKeyAlias
checking and abort client connections in case HostKeyAlias is not
present in the target host certificate.

That means we must provide a different HostKeyAlias for each of the
target hosts.  The easiest way to achieve that is to add the option to
the ssh command line for console attachment.

Change-Id: Ia76140e25f90ae82bc8e3c6e9e370f6c0d643686
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>